requirements.txt
anthropic>=0.40.0
aiohttp>=3.9.0
python-dotenv>=1.0.0
twilio>=8.0.0
flask>=3.0.0
rich>=13.0.0
